--- Artemis Project/Assets/Scripts/Question.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: af8294896b5abb04dbe11fad1cc267e3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
